chore(): Move file to follow directory structure
