+++ lib/activeadmin_addons/support/set_datepicker.rb
